LayoutTests/fast/dom/SelectorAPI/caseTag.html

<html>
<head>
<link rel="stylesheet" href="../../js/resources/js-test-style.css">
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
    <p id="description"></p>
    <div id="console"></div>
    <div style="display: none">
    <span id="lower1">lower</span><br>
    <P id="UPPER1">UPPER</P><br>
    </div>
    <script>
        shouldBe("document.querySelector('div SPAN').textContent", "'lower'");
        shouldBe("document.querySelector('div p').textContent", "'UPPER'");

        shouldBeTrue("document.getElementById('lower1').webkitMatchesSelector('div SPAN')");
        shouldBeTrue("document.getElementById('UPPER1').webkitMatchesSelector('div p')");

        var successfullyParsed = true;
    </script>
    <script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
