LayoutTests/fast/dom/SelectorAPI/caseTag-expected.txt

PASS document.querySelector('div SPAN').textContent is 'lower'
PASS document.querySelector('div p').textContent is 'UPPER'
PASS document.getElementById('lower1').webkitMatchesSelector('div SPAN') is true
PASS document.getElementById('UPPER1').webkitMatchesSelector('div p') is true
PASS successfullyParsed is true

TEST COMPLETE
